led_panel_driver: fix incorrect point_to_index
The typedef for point_to_index expects driver as the first parameter.
For some reason -Werror was not turned on even though ESP says it is
on by default. Fix this function, add explicit error checks to cmake
definition to avoid this issue in the future. Also make points and
sizes completely consistent and unsigned throughout.