bitmap: add bitmap draw api
Define a bitmap as raw GBR pixels and draw them directly on the surface.
This is not entirely useful but it looks cool. We need a BGR2HSV tool
in order to control the brightness. With all 512 pixels let my power
supply hits nearly 5 amps current draw.

Also add an idle loop to main in case an example returns early. This
avoids a WDT reset.