led_panel_driver: implement dedicated render task
Add a timer to maintain a stable frame rate. Now that we have
non-blocking writes and a task-safe surface, we can remove the delay
from the driver. This allows the client to manage their own frame rate.

Initial testing shows a much smoother rendering but I don't have numbers
to prove this.